refactor: consolidate parser in a folder
