+++ src/server/features/completion/sassdoc-completion.ts
@@ -5,7 +5,7 @@ import {
 	InsertTextFormat,
 } from "vscode-languageserver-types";
 import { sassDocAnnotations } from "../../sassdoc-annotations";
-import { getLinesFromText } from "../../utils/document";
+import { getLinesFromText } from "../../utils/string";
 import type { CompletionContext } from "./completion-context";
 
 export function doSassDocCompletion(
