+++ src/server/parser/parser.ts
@@ -1,4 +1,4 @@
-import { Parameter, parseString } from "scss-sassdoc-parser";
+import { parseString } from "scss-sassdoc-parser";
 import type { ParseResult } from "scss-sassdoc-parser";
 import {
 	Position,
@@ -8,190 +8,15 @@ import {
 } from "vscode-css-languageservice";
 import type { TextDocument } from "vscode-languageserver-textdocument";
 import { URI } from "vscode-uri";
+import { sassBuiltInModuleNames } from "../sass-built-in-modules";
+import { fileExists } from "../utils/fs";
+import { asDollarlessVariable, getLinesFromText } from "../utils/string";
+import { getNodeAtOffset, getParentNodeByType } from "./ast";
+import { buildDocumentContext } from "./document";
 import { getLanguageService } from "./language-service";
-import { sassBuiltInModuleNames } from "./sass-built-in-modules";
-import { buildDocumentContext, getLinesFromText } from "./utils/document";
-import { fileExists, realPath } from "./utils/fs";
-import { asDollarlessVariable } from "./utils/string";
-
-/**
- * Must be synced with https://github.com/microsoft/vscode-css-languageservice/blob/main/src/parser/cssNodes.ts
- * when upgrading vscode-css-languageservice.
- */
-export enum NodeType {
-	Undefined,
-	Identifier,
-	Stylesheet,
-	Ruleset,
-	Selector,
-	SimpleSelector,
-	SelectorInterpolation,
-	SelectorCombinator,
-	SelectorCombinatorParent,
-	SelectorCombinatorSibling,
-	SelectorCombinatorAllSiblings,
-	SelectorCombinatorShadowPiercingDescendant,
-	Page,
-	PageBoxMarginBox,
-	ClassSelector,
-	IdentifierSelector,
-	ElementNameSelector,
-	PseudoSelector,
-	AttributeSelector,
-	Declaration,
-	Declarations,
-	Property,
-	Expression,
-	BinaryExpression,
-	Term,
-	Operator,
-	Value,
-	StringLiteral,
-	URILiteral,
-	EscapedValue,
-	Function,
-	NumericValue,
-	HexColorValue,
-	RatioValue,
-	MixinDeclaration,
-	MixinReference,
-	VariableName,
-	VariableDeclaration,
-	Prio,
-	Interpolation,
-	NestedProperties,
-	ExtendsReference,
-	SelectorPlaceholder,
-	Debug,
-	If,
-	Else,
-	For,
-	Each,
-	While,
-	MixinContentReference,
-	MixinContentDeclaration,
-	Media,
-	Keyframe,
-	FontFace,
-	Import,
-	Namespace,
-	Invocation,
-	FunctionDeclaration,
-	ReturnStatement,
-	MediaQuery,
-	MediaCondition,
-	MediaFeature,
-	FunctionParameter,
-	FunctionArgument,
-	KeyframeSelector,
-	ViewPort,
-	Document,
-	AtApplyRule,
-	CustomPropertyDeclaration,
-	CustomPropertySet,
-	ListEntry,
-	Supports,
-	SupportsCondition,
-	NamespacePrefix,
-	GridLine,
-	Plugin,
-	UnknownAtRule,
-	Use,
-	ModuleConfiguration,
-	Forward,
-	ForwardVisibility,
-	Module,
-	UnicodeRange,
-}
-
-export interface INode {
-	type: NodeType;
-	offset: number;
-	length: number;
-	end: number;
-	accept: (node: (node: INode) => boolean) => boolean;
-	getName: () => string;
-	getValue: () => INode;
-	getDefaultValue: () => INode;
-	getText: () => string;
-	getParameters: () => INode;
-	getIdentifier: () => INode;
-	getParent: () => INode;
-	getChildren: () => INode[];
-	getChild: (index: number) => INode;
-	getSelectors: () => INode;
-}
-
-export interface ScssSymbol {
-	kind: SymbolKind;
-	name: string;
-	sassdoc?: ParseResult;
-	position: Position;
-	offset: number;
-}
-
-export interface ScssVariable extends ScssSymbol {
-	mixin?: string;
-	value: string | null;
-}
-
-export interface ScssParameter
-	extends Omit<ScssVariable, "kind" | "position" | "sassdoc"> {
-	sassdoc?: Parameter;
-}
-
-export interface ScssMixin extends ScssSymbol {
-	parameters: ScssParameter[];
-}
-
-export type ScssFunction = ScssMixin;
-
-export interface ScssLink {
-	link: DocumentLink;
-}
-
-export interface ScssUse extends ScssLink {
-	namespace?: string;
-	/** Indicates whether the namespace is different from the file name. */
-	isAliased: boolean;
-}
-
-export interface ScssForward extends ScssLink {
-	hide: string[];
-	prefix?: string;
-}
-
-export interface ScssImport extends ScssLink {
-	css: boolean;
-	dynamic: boolean;
-}
-
-export interface IScssSymbols {
-	imports: Map<string, ScssImport>;
-	uses: Map<string, ScssUse>;
-	forwards: Map<string, ScssForward>;
-	variables: Map<string, ScssVariable>;
-	mixins: Map<string, ScssMixin>;
-	functions: Map<string, ScssFunction>;
-}
-
-export interface IScssDocument extends TextDocument, IScssSymbols {
-	textDocument: TextDocument;
-	ast: INode;
-	filePath: string;
-	/**
-	 * The last part of the URI, including extension.
-	 * For instance, given the URI `file:///home/test.scss`,
-	 * the fileName is `test.scss`.
-	 */
-	fileName: string;
-	/** Find and cache the real path (as opposed to symlinked) */
-	getRealPath: () => Promise<string | null>;
-	getLinks: (options?: { forwards: boolean }) => ScssLink[];
-	getSymbols: () => ScssSymbol[];
-	getNodeAt: (offset: number) => INode | null;
-	getNodeRange: (node: INode) => Range;
-}
+import { INode, NodeType } from "./node";
+import { ScssDocument } from "./scss-document";
+import type { IScssSymbols } from "./scss-symbol";
 
 export const reModuleAtRule = /@(?:use|forward|import)/;
 export const reUse = /@use ["'|](?<url>.+)["'|](?: as (?<namespace>\*|\w+))?;/;
@@ -507,195 +332,4 @@ function getMethodParameters(
 				sassdoc: docs,
 			};
 		});
-}
-
-/**
- * Get Node by offset position.
- */
-export function getNodeAtOffset(
-	parsedDocument: INode,
-	posOffset: number | null,
-): INode | null {
-	let candidate: INode | null = null;
-
-	parsedDocument.accept((node) => {
-		if (node.offset === -1 && node.length === -1) {
-			return true;
-		}
-
-		if (
-			posOffset !== null &&
-			node.offset <= posOffset &&
-			node.end >= posOffset
-		) {
-			if (!candidate) {
-				candidate = node;
-			} else if (node.length <= candidate.length) {
-				candidate = node;
-			}
-
-			return true;
-		}
-
-		return false;
-	});
-
-	return candidate;
-}
-
-/**
- * Returns the parent Node of the specified type.
- */
-export function getParentNodeByType(
-	node: INode | null,
-	type: NodeType,
-): INode | null {
-	if (node === null) {
-		return null;
-	}
-
-	node = node.getParent();
-
-	while (node.type !== type) {
-		if (node.type === NodeType.Stylesheet) {
-			return null;
-		}
-
-		node = node.getParent();
-	}
-
-	return node;
-}
-
-export class ScssDocument implements IScssDocument {
-	public textDocument: TextDocument;
-	public ast: INode;
-	public fileName: string;
-
-	public imports: Map<string, ScssImport> = new Map();
-	public uses: Map<string, ScssUse> = new Map();
-	public forwards: Map<string, ScssForward> = new Map();
-	public variables: Map<string, ScssVariable> = new Map();
-	public mixins: Map<string, ScssMixin> = new Map();
-	public functions: Map<string, ScssFunction> = new Map();
-
-	private _realPath: string | null = null;
-
-	constructor(document: TextDocument, symbols: IScssSymbols, ast?: INode) {
-		this.textDocument = document;
-		this.ast = ast ?? (ls.parseStylesheet(document) as INode);
-		this.imports = symbols.imports;
-		this.uses = symbols.uses;
-		this.forwards = symbols.forwards;
-		this.variables = symbols.variables;
-		this.mixins = symbols.mixins;
-		this.functions = symbols.functions;
-		this.fileName = this.getFileName();
-	}
-
-	public get uri(): string {
-		return this.textDocument.uri;
-	}
-
-	public get filePath(): string {
-		return URI.parse(this.textDocument.uri).fsPath;
-	}
-
-	public async getRealPath(): Promise<string | null> {
-		if (this._realPath) {
-			return this._realPath;
-		}
-
-		try {
-			const path = await realPath(this.filePath);
-			this._realPath = path;
-		} catch {
-			// Do nothing
-		}
-
-		return this._realPath;
-	}
-
-	private getFileName(): string {
-		const uri = this.textDocument.uri;
-		const lastSlash = uri.lastIndexOf("/");
-		return lastSlash === -1 ? uri : uri.slice(Math.max(0, lastSlash + 1));
-	}
-
-	public get languageId(): string {
-		return this.textDocument.languageId;
-	}
-
-	public get version(): number {
-		return this.textDocument.version;
-	}
-
-	public getText(range?: Range): string {
-		return this.textDocument.getText(range);
-	}
-
-	public getNodeAt(offset: number): INode | null {
-		return getNodeAtOffset(this.ast, offset);
-	}
-
-	public getNodeRange(node: INode): Range {
-		return Range.create(
-			this.textDocument.positionAt(node.offset),
-			this.textDocument.positionAt(node.end),
-		);
-	}
-
-	public positionAt(offset: number): Position {
-		return this.textDocument.positionAt(offset);
-	}
-
-	public offsetAt(position: Position): number {
-		return this.textDocument.offsetAt(position);
-	}
-
-	public get lineCount(): number {
-		return this.textDocument.lineCount;
-	}
-
-	public getLines(): string[] {
-		return getLinesFromText(this.textDocument.getText());
-	}
-
-	public getSymbols(): ScssSymbol[] {
-		const symbols: ScssSymbol[] = [];
-
-		for (const variable of this.variables.values()) {
-			symbols.push(variable);
-		}
-
-		for (const mixin of this.mixins.values()) {
-			symbols.push(mixin);
-		}
-
-		for (const func of this.functions.values()) {
-			symbols.push(func);
-		}
-
-		return symbols;
-	}
-
-	public getLinks(options = { forwards: true }): ScssLink[] {
-		const links: ScssLink[] = [];
-
-		for (const imp of this.imports.values()) {
-			links.push(imp);
-		}
-
-		for (const use of this.uses.values()) {
-			links.push(use);
-		}
-
-		if (options.forwards) {
-			for (const forward of this.forwards.values()) {
-				links.push(forward);
-			}
-		}
-
-		return links;
-	}
 }
